test: update feed editor controller tests
