--- lib/pages/create_feed/bindings/create_feed_binding.dart
@@ -1,9 +1,0 @@
-import 'package:get/get.dart';
-import 'package:hoot/pages/create_feed/controllers/create_feed_controller.dart';
-
-class CreateFeedBinding extends Bindings {
-  @override
-  void dependencies() {
-    Get.lazyPut(() => CreateFeedController());
-  }
-}
