--- lib/pages/edit_feed/bindings/edit_feed_binding.dart
@@ -1,9 +1,0 @@
-import 'package:get/get.dart';
-import 'package:hoot/pages/edit_feed/controllers/edit_feed_controller.dart';
-
-class EditFeedBinding extends Bindings {
-  @override
-  void dependencies() {
-    Get.lazyPut(() => EditFeedController());
-  }
-}
