+++ lib/pages/feed_editor/bindings/feed_editor_binding.dart
@@ -1,0 +1,9 @@
+import 'package:get/get.dart';
+import 'package:hoot/pages/feed_editor/controllers/feed_editor_controller.dart';
+
+class FeedEditorBinding extends Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut(() => FeedEditorController());
+  }
+}
